+++ src/main/java/org/bedu/atko/entity/Category.java
@@ -1,16 +1,14 @@
 package org.bedu.atko.entity;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "category")
 public class Category {
