+++ src/main/java/org/bedu/atko/entity/Reviews.java
@@ -1,15 +1,13 @@
 package org.bedu.atko.entity;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "reviews")
 public class Reviews {
